refactor: rename and move slices and loaders, divide search-result component
