+++ task-1/src/features/index.ts
@@ -1,3 +1,1 @@
-export * from './stateInterface';
 export * from './items-selector';
-export * from './pageItemsSlice';
