+++ task-1/src/features/items-selector/index.ts
@@ -1,2 +1,1 @@
-export * from './itemsSelectorSlice';
 export * from './itemsSelector';
