+++ task-1/src/loaders/details-loader.ts
