+++ task-1/src/loaders/index.ts
@@ -1,0 +1,2 @@
+export * from './details-loader';
+export * from './results-loader';
