+++ task-1/src/slices/index.ts
@@ -1,0 +1,3 @@
+export * from './state-interface';
+export * from './items-selector-slice';
+export * from './page-items-slice';
