+++ task-1/src/slices/state-interface.ts
