--- task-1/src/utils/index.ts
@@ -1,2 +1,0 @@
-export * from './detailsLoader';
-export * from './resultsLoader';
