+++ task-1/src/views/search-page/components/index.ts
@@ -2,3 +2,4 @@ export * from './search-button';
 export * from './search-input';
 export * from './terms-list';
 export * from './search-results';
+export * from './result-card';
