+++ task-1/src/views/search-page/components/result-card/index.ts
@@ -1,0 +1,1 @@
+export * from './result-card';
