+++ task-1/src/views/search-page/components/search-results/search-results.tsx
@@ -3,19 +3,11 @@ import {
   CharData,
   AllCharsData,
 } from '../../../../services';
-import { Link } from 'react-router-dom';
-import { ItemsSelector } from 'src/features';
+import { ResultCard } from '..';
 
 interface SearchResultsProps {
   searchResults?: AllCharsData;
   className?: string;
-}
-
-interface SectionDescription {
-  name: string;
-  type: string;
-  text: string;
-  _id: number;
 }
 
 function getDescription(charData: CharData): string[] {
@@ -31,59 +23,31 @@ function getDescription(charData: CharData): string[] {
   );
 }
 
-function getSection(
-  description: SectionDescription,
-  data: CharData
-): JSX.Element {
-  const { name, type, text, _id } = { ...description };
-
-  const onClick: React.MouseEventHandler<
-    HTMLAnchorElement
-  > = (event) => {
-    if (
-      event.target &&
-      (event.target as HTMLElement).closest(
-        '.items-selector'
-      )
-    ) {
-      event.preventDefault();
-    }
-  };
-
-  return (
-    <Link to={`characters/${_id}`} onClick={onClick}>
-      <section>
-        <ItemsSelector data={data}></ItemsSelector>
-        <h3>{name}</h3>
-        <div>
-          <h4>{type}</h4>
-          <p>{text}</p>
-        </div>
-      </section>
-    </Link>
-  );
+function getResultElements(
+  props: SearchResultsProps
+): JSX.Element[] {
+  const data = props.searchResults?.data;
+  const cardsArr: JSX.Element[] =
+    data?.map((res) => {
+      const description = getDescription(res);
+      return ResultCard({
+        _id: res._id,
+        name: res.name,
+        type: description[0],
+        text: description[1],
+        data: res,
+      });
+    }) ?? [];
+  return cardsArr;
 }
 
 export const SearchResults = (
   props: SearchResultsProps
 ) => {
-  const cardsArr = useMemo(() => {
-    const data = props.searchResults?.data;
-    const cardsArr: JSX.Element[] =
-      data?.map((res) => {
-        const description = getDescription(res);
-        return getSection(
-          {
-            _id: res._id,
-            name: res.name,
-            type: description[0],
-            text: description[1],
-          },
-          res
-        );
-      }) ?? [];
-    return cardsArr;
-  }, [props]);
+  const cardsArr = useMemo(
+    () => getResultElements(props),
+    [props]
+  );
 
   return (
     <div className={props.className}>{...cardsArr}</div>
